feat: gpt-4 as a judge for prompt answers
Add an optional `final_answer_should` field to test specs that, when
set, allows test authors to define a set of constraints used to
determine the correctness of the final natural language response
returned by the model for a prompt.

When constraints are provided, the model's final response must satisfy
all of them in order for a test to pass.

Constraints are written in natural language and are given, along with
the model's response, to OpenAI's `gpt-4-turbo-preview` model to
act as a judge.

When an answer is judged as incorrect, a brief summary supporting the
conclusion is included with the failed assertion to provide more
context.

e.g.

```
  {
    "description": "Calls a function with a single parameter and provided input",
    "categories": [
      "basic"
    ],
    "prompt": "Call funcA with 1 and respond with the result",
    "available_functions": [
      {
        "name": "funcA",
        "description": "Performs funcA",
        "parameters": {
          "type": "object",
          "properties": {
            "param1": {
              "type": "string",
              "description": "Param 1"
            }
          }
        }
      }
    ],
    "expected_function_calls": [
      {
        "name": "funcA",
        "arguments": {
          "param1": "1"
        },
        "result": "This is the output of funcA(1)"
      },
      {
        "finish_reason": "stop"
      }
    ],
    "final_answer_should": "The answer should indicate that the result of calling funcA with 1 is \"This is the output of funcA(1)\""
  }
```

Signed-off-by: Nick Hale <4175918+njhale@users.noreply.github.com>