Allow any order for chained test
Sometimes, GPT4 figures out that it can call all of the "leaf node"
functions in one go before calling the functions that rely on their
output. So, it is making all the correct calls, but actually a bit more
efficiently. To fix this, we should just allow all the functions in any
order and rely on the inputs being correct for the expected function
calls.

Signed-off-by: Craig Jellick <craig@acorn.io>