Fix: Make semantic_negative ruling more lenient
In this test, the time cannot be determined. The point of the test is to
make sure the model doesn't call a non-existant time function and
doesn't hallucinate a time. Sometimes, GPT4 will suggest a way to
determine the time. It is not actually giving the time, just telling the
user how they can figure it out on their own. This should be allowed.

Signed-off-by: Craig Jellick <craig@acorn.io>