Fix: Make basic_negative ruling more lenient
Give this answer: "It seems there was a misunderstanding in the input provided.
The function `funcA` requires an integer parameter named `param1`.
However, you provided a string value 'foo'. Please provide an integer value
to proceed with calling `funcA`."

I got a FAIL ruling from the judge, but that answer should PASS. This
addresses the issue by loosening the judgement condition a little.

Signed-off-by: Craig Jellick <craig@acorn.io>